more cleaning up - moved / created simulation.py
