+++ src/globals/globals.py
@@ -1,3 +1,4 @@
 ENGINE_COLLECTION = None
 PORTFOLIO = None
 LOCALSTARTTIME = None
+THREADENABLED = None
