+++ templates/aapl.html
@@ -16,7 +16,7 @@
     <script src="http://code.highcharts.com/highcharts-more.js"></script>
     <script src="http://code.highcharts.com/modules/exporting.js"></script>
 
-    <script src="../apis.js"></script>
+<!--    <script src="../apis.js"></script>-->
     <script>
         $(document).ready(function() {
            $('form').submit(function(event) {
